# rewrite main gui to be non-destructive and faster