--- a/maps/biter_battles_v2/ai.lua
+++ b/maps/biter_battles_v2/ai.lua
@@ -490,19 +490,15 @@ end
 
 function Public.reanimate_units()
 	-- This FIFOs can be accessed by force indices.
-	for force, id in pairs(global.dead_units) do
+	for _, id in pairs(global.dead_units) do
 		-- Check for each side if there are any biters to reanimate.
-		if fifo.empty(id) then
-			goto reanim_units_cont
+		if not fifo.empty(id) then
+			-- Balance amount of unit creation requests to get rid off
+			-- excess stored in memory.
+			local cycles = fifo.length(id) / global.reanim_balancer
+			cycles = math.floor(cycles) + 1
+			_reanimate_units(id, cycles)
 		end
-
-		-- Balance amount of unit creation requests to get rid off
-		-- excess stored in memory.
-		local cycles = fifo.length(id) / global.reanim_balancer
-		cycles = math.floor(cycles) + 1
-		_reanimate_units(id, cycles)
-
-		::reanim_units_cont::
 	end
 end