--- a/maps/biter_battles_v2/ai.lua
+++ b/maps/biter_battles_v2/ai.lua
@@ -492,7 +492,7 @@ function Public.reanimate_units()
 	-- This FIFOs can be accessed by force indices.
 	for _, id in pairs(global.dead_units) do
 		-- Check for each side if there are any biters to reanimate.
-		if not fifo.empty(id) then
+		if fifo.empty(id) then
 			goto reanim_units_cont
 		end