--- a/maps/biter_battles_v2/difficulty_vote.lua
+++ b/maps/biter_battles_v2/difficulty_vote.lua
@@ -1,92 +1,66 @@
-local bb_config = require("maps.biter_battles_v2.config")
-local ai = require("maps.biter_battles_v2.ai")
-local event = require("utils.event")
-local Server = require("utils.server")
-local Tables = require("maps.biter_battles_v2.tables")
-require("utils/gui_styles")
+local bb_config = require "maps.biter_battles_v2.config"
+local ai = require "maps.biter_battles_v2.ai"
+local event = require 'utils.event'
+local Server = require 'utils.server'
+local Tables = require "maps.biter_battles_v2.tables"
+require 'utils/gui_styles'
 
 local difficulties = Tables.difficulties
 
 local function create_difficulty_gui_for_player(player)
-	local diff_gui = player.gui.top.add({
-		type = "sprite-button",
-		caption = difficulties[global.difficulty_vote_index].name,
-		tooltip = "",
-		name = "difficulty_gui",
-	})
+	local diff_gui = player.gui.top.add{ type = "sprite-button", caption = difficulties[global.difficulty_vote_index].name, tooltip = "", name = "difficulty_gui" }
 	diff_gui.style.font = "heading-2"
 	diff_gui.style.font_color = difficulties[global.difficulty_vote_index].print_color
-	element_style({ element = diff_gui, x = 114, y = 38, pad = -2 })
+	element_style({element = diff_gui, x = 114, y = 38, pad = -2})
 end
 
 local function update_difficulty_gui_for_player(player, value)
-	if not player.gui.top.difficulty_gui then
-		create_difficulty_gui_for_player(player)
-	end
+	if not player.gui.top.difficulty_gui then create_difficulty_gui_for_player(player) end
 	local diff_gui = player.gui.top.difficulty_gui
-	local diff_name = difficulties[global.difficulty_vote_index].name
-	local str = table.concat({ "Global map difficulty is ", diff_name, ". Mutagen has ", value, "% effectiveness." })
+	local str = table.concat({"Global map difficulty is ",
+		difficulties[global.difficulty_vote_index].name,
+		". Mutagen has ", value,
+		"% effectiveness."})
 	diff_gui.tooltip = str
-	diff_gui.caption = diff_name
 end
 
 local function update_difficulty_gui_for_all()
-	local value = math.floor(global.difficulty_vote_value * 100)
+	local value = math.floor(global.difficulty_vote_value*100)
 	for _, player in pairs(game.connected_players) do
 		update_difficulty_gui_for_player(player, value)
 	end
 end
 
 local function poll_difficulty(player)
-	if player.gui.center["difficulty_poll"] then
-		player.gui.center["difficulty_poll"].destroy()
-		return
-	end
-
+	if player.gui.center["difficulty_poll"] then player.gui.center["difficulty_poll"].destroy() return end
+	
 	if global.bb_settings.only_admins_vote or global.tournament_mode then
-		if not player.admin then
-			return
-		end
+		if not player.admin then return end
 	end
-
+	
 	local tick = game.ticks_played
 	if tick > global.difficulty_votes_timeout then
 		if player.online_time ~= 0 then
 			local t = math.abs(math.floor((global.difficulty_votes_timeout - tick) / 3600))
 			local str = "Votes have closed " .. t
 			str = str .. " minute"
-			if t > 1 then
-				str = str .. "s"
-			end
+			if t > 1 then str = str .. "s" end
 			str = str .. " ago."
 			player.print(str)
 		end
-		return
+		return 
 	end
-
-	local frame = player.gui.center.add({
-		type = "frame",
-		caption = "Vote global difficulty:",
-		name = "difficulty_poll",
-		direction = "vertical",
-	})
+	
+	local frame = player.gui.center.add { type = "frame", caption = "Vote global difficulty:", name = "difficulty_poll", direction = "vertical" }
 	for key, _ in pairs(difficulties) do
-		local b = frame.add({
-			type = "button",
-			name = tostring(key),
-			caption = difficulties[key].name .. " (" .. difficulties[key].str .. ")",
-		})
+		local b = frame.add({type = "button", name = tostring(key), caption = difficulties[key].name .. " (" .. difficulties[key].str .. ")"})
 		b.style.font_color = difficulties[key].color
 		b.style.font = "heading-2"
 		b.style.minimal_width = 180
 	end
-	local b = frame.add({ type = "label", caption = "- - - - - - - - - - - - - - - - - - - -" })
-	local b = frame.add({
-		type = "button",
-		name = "close",
-		caption = "Close (" .. math.floor((global.difficulty_votes_timeout - tick) / 3600) .. " minutes left)",
-	})
-	b.style.font_color = { r = 0.66, g = 0.0, b = 0.66 }
+	local b = frame.add({type = "label", caption = "- - - - - - - - - - - - - - - - - - - -"})
+	local b = frame.add({type = "button", name = "close", caption = "Close (" .. math.floor((global.difficulty_votes_timeout - tick) / 3600) .. " minutes left)"})
+	b.style.font_color = {r=0.66, g=0.0, b=0.66}
 	b.style.font = "heading-3"
 	b.style.minimal_width = 96
 end
@@ -101,104 +75,65 @@ local function set_difficulty()
 		a[c] = d
 		vote_count = vote_count + 1
 	end
-	if vote_count == 0 then
-		return
-	end
-	v = math.floor(vote_count / 2) + 1
+	if vote_count == 0 then return end
+	v= math.floor(vote_count/2)+1
 	table.sort(a)
 	local new_index = a[v]
 	if global.difficulty_vote_index ~= new_index then
-		local message = table.concat({
-			">> Map difficulty has changed to ",
-			difficulties[new_index].name,
-			" difficulty!",
-		})
+		local message = table.concat({">> Map difficulty has changed to ", difficulties[new_index].name, " difficulty!"})
 		game.print(message, difficulties[new_index].print_color)
 		Server.to_discord_embed(message)
 	end
-	global.difficulty_vote_index = new_index
-	global.difficulty_vote_value = difficulties[new_index].value
-	ai.reset_evo()
+	 global.difficulty_vote_index = new_index
+	 global.difficulty_vote_value = difficulties[new_index].value
+	 ai.reset_evo()
 end
 
 local function on_player_joined_game(event)
-	if not global.difficulty_vote_value then
-		global.difficulty_vote_value = 1
-	end
-	if not global.difficulty_vote_index then
-		global.difficulty_vote_index = 4
-	end
-	if not global.difficulty_player_votes then
-		global.difficulty_player_votes = {}
-	end
-
+	if not global.difficulty_vote_value then global.difficulty_vote_value = 1 end
+	if not global.difficulty_vote_index then global.difficulty_vote_index = 4 end
+	if not global.difficulty_player_votes then global.difficulty_player_votes = {} end
+	
 	local player = game.players[event.player_index]
 	if game.ticks_played < global.difficulty_votes_timeout then
 		if not global.difficulty_player_votes[player.name] then
 			if global.bb_settings.only_admins_vote or global.tournament_mode then
-				if player.admin then
-					poll_difficulty(player)
-				end
+				if player.admin then poll_difficulty(player) end
 			end
 		end
 	else
-		if player.gui.center["difficulty_poll"] then
-			player.gui.center["difficulty_poll"].destroy()
-		end
+		if player.gui.center["difficulty_poll"] then player.gui.center["difficulty_poll"].destroy() end
 	end
-
+	
 	update_difficulty_gui_for_all()
 end
 
 local function on_player_left_game(event)
-	if game.ticks_played > global.difficulty_votes_timeout then
-		return
-	end
+	if game.ticks_played > global.difficulty_votes_timeout then return end
 	local player = game.players[event.player_index]
-	if not global.difficulty_player_votes[player.name] then
-		return
-	end
+	if not global.difficulty_player_votes[player.name] then return end
 	global.difficulty_player_votes[player.name] = nil
 	set_difficulty()
 end
 
 local function on_gui_click(event)
-	if not event then
-		return
-	end
-	if not event.element then
-		return
-	end
-	if not event.element.valid then
-		return
-	end
+	if not event then return end
+	if not event.element then return end
+	if not event.element.valid then return end
 	local player = game.players[event.element.player_index]
 	if event.element.name == "difficulty_gui" then
 		poll_difficulty(player)
 		return
 	end
-	if event.element.type ~= "button" then
-		return
-	end
-	if event.element.parent.name ~= "difficulty_poll" then
-		return
-	end
-	if event.element.name == "close" then
-		event.element.parent.destroy()
-		return
-	end
-	if game.ticks_played > global.difficulty_votes_timeout then
-		event.element.parent.destroy()
-		return
-	end
+	if event.element.type ~= "button" then return end
+	if event.element.parent.name ~= "difficulty_poll" then return end
+	if event.element.name == "close" then event.element.parent.destroy() return end
+	if game.ticks_played > global.difficulty_votes_timeout then event.element.parent.destroy() return end
 	local i = tonumber(event.element.name)
-
+	
 	if global.bb_settings.only_admins_vote or global.tournament_mode then
 		if player.admin then
-			game.print(
-				player.name .. " has voted for " .. difficulties[i].name .. " difficulty!",
-				difficulties[i].print_color
-			)
+			game.print(player.name .. " has voted for " .. difficulties[i].name .. " difficulty!", difficulties[i].print_color)
 			global.difficulty_player_votes[player.name] = i
 			set_difficulty()
 			update_difficulty_gui_for_all()
@@ -207,30 +142,28 @@ local function on_gui_click(event)
 		return
 	end
 
-	if player.spectator then
-		player.print("spectators can't vote for difficulty")
+    if player.spectator then
+        player.print("spectators can't vote for difficulty")
 		event.element.parent.destroy()
-		return
-	end
+        return
+    end
 
 	if game.tick - global.spectator_rejoin_delay[player.name] < 3600 then
-		player.print(
-			"Not ready to vote. Please wait "
-				.. 60 - (math.floor((game.tick - global.spectator_rejoin_delay[player.name]) / 60))
-				.. " seconds.",
-			{ r = 0.98, g = 0.66, b = 0.22 }
-		)
+        player.print(
+            "Not ready to vote. Please wait " .. 60-(math.floor((game.tick - global.spectator_rejoin_delay[player.name])/60)) .. " seconds.",
+            {r = 0.98, g = 0.66, b = 0.22}
+        )
 		event.element.parent.destroy()
-		return
-	end
-
+        return
+    end
+	
 	game.print(player.name .. " has voted for " .. difficulties[i].name .. " difficulty!", difficulties[i].print_color)
 	global.difficulty_player_votes[player.name] = i
 	set_difficulty()
 	update_difficulty_gui_for_all()
 	event.element.parent.destroy()
 end
-
+	
 event.add(defines.events.on_gui_click, on_gui_click)
 event.add(defines.events.on_player_left_game, on_player_left_game)
 event.add(defines.events.on_player_joined_game, on_player_joined_game)