… maps/biter_battles_v2/main.lua
… maps/biter_battles_v2/team_manager.lua