### Brief description of the changes:
Instead of tearing down the gui and then building it from scratch, it is relatively simple to update the values of each element. (we can use .visible instead of .destroy() to get rid of things)

I think the last place to squeeze some performance out is to somehow create a single gui once for everyone, and then only set .visible flags per person in the update_main_gui function. but I think I will give up here since I have wasted enough time on this.

this touches a lot of code but with a single person in lobby the average time for .refresh has gone from about 4ms to 0.8ms... but with maximum spikes changing from 10.6 ms to 3ms

A truly laughable improvement.... :'(

but maybe it will be more improved with more people.....

anyway, I didn't test this in multiplayer or anything so there's probably little bits that are broken - I'm not super well versed in what causes desyncs.

I am also surprised at how long reanimate_units takes. I wonder if that should be moved into main.lua or something because it's a bit of a resource hog even at 0 evo.

also worth noting, it can be very hard to profile code because of the on_generate_chunk code... it seems that is not very friendly.....

also the difficulty option meter moved... can be easily fixed but I am taking a break from this code.


### Tested Changes:
- [ ] I've tested the changes locally or with people.
- [ ] I've not tested the changes.