@AwesomePatrol 2021-12-14
There are some issues left (missing science button in spectate):
```
  51.829 Script @/home/patrol/.factorio/temp/currently-playing/utils/event_core.lua:48: LuaGuiElement API call when LuaGuiElement was invalid.
  55.275 Script @/home/patrol/.factorio/temp/currently-playing/utils/event_core.lua:48: ...rio/temp/currently-playing/maps/biter_battles_v2/gui.lua:281: attempt to index field 'bb_science_frame' (a nil value)
  65.625 Script @/home/patrol/.factorio/temp/currently-playing/utils/event_core.lua:48: LuaGuiElement API call when LuaGuiElement was invalid.
  ```