@developer-8 2021-12-18
After changing the difficulty, the difficulty button does not change