@Ragnarok77-factorio 2021-12-30
for @XVhc6A  : 

=>I opened gui with fish, closed it, had this error : 
utils/event_core.lua:48: LuaGuiElement API call when LuaGuiElement was invalid

=>When I host it, I see the buttons in black instead of previous color (harder to read now), example here of the screen before and after your change : 
![image](https://user-images.githubusercontent.com/95619848/147786384-9e1cf1fb-4a51-4ca9-91e0-a19b6f72e99d.png)

=>When I clicked auto join, got this error : 
/event_core.lua:48: ...rio/temp/currently-playing/maps/biter_battles_v2/gui.lua:513: attempt to index field 'bb_main_gui' (a nil value)

=>When I feed red science in training mode, no threat is increased/shown on gui, or even not training mode, it shows 0 (maybe just an illusion)